adding ruteo por resposabilidad
