+++ routes/index.js
@@ -1,0 +1,9 @@
+const productsRouter = require('./products.router');
+// const usersRouter = require('./users.router');
+
+function routerApi(app) {
+  app.use('/products', productsRouter);
+  // app.use('/users', usersRouter);
+}
+
+module.exports = routerApi;
