+++ routes/products.router.js
@@ -1,0 +1,33 @@
+const express = require('express');
+const { faker } = require('@faker-js/faker');
+
+const router = express.Router();
+
+router.get('/', (req, res) => {
+  const products = [];
+  const { size } = req.query;
+  const limit = size || 10;
+  for (let index = 0; index < 100; index++) {
+    products.push({
+      name: faker.commerce.productName(),
+      price: parseInt(faker.commerce.price(), 10),
+      image: faker.image.imageUrl(),
+    });
+  }
+  res.json(products);
+});
+
+router.get('/filter', (req, res) => {
+  res.send('Yo soy un filter');
+})
+
+router.get('/:id', (req, res) => {
+  const { id }= req.params;
+  res.json({
+    id,
+    name: 'Producto 2',
+    price: 2000
+  });
+});
+
+module.exports = router;
