+++ routes/users.router.js
@@ -1,0 +1,15 @@
+const express = require('express');
+
+const router = express.Router();
+
+router.get('/users', (req, res) => {
+  const { limit, offset } = req.query;
+  if (limit && offset) {
+    res.json({
+      limit,
+      offset
+    });
+  } else {
+    res.send('No hay parametros');
+  }
+});
